+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - variable commands in settings
 
+### Fixed
+
+- fixed windows powershell bug when using &&
+
 ## [0.1.0] - 2025-04-09
 
 ### Added